+++ apps/extractAudioFromInsv/extractAudioFromInsv.py
@@ -1,14 +1,14 @@
 import os
 import subprocess
 from concurrent.futures import ThreadPoolExecutor
-from apps.FileNameFunctions import increment_filename
+from apps.misc.FileNameFunctions import increment_filename
 
 
 # THIS IS A GUI PROGRAM! PLEASE RUN FILE TO MAKE GUI!
 
 
 # Function to extract audio from video files
-def extract_audio(video_path:str, output_folder:str) -> str:
+def extract_audio(video_path: str, output_folder: str) -> str:
     # Define the output audio file path
     audio_path = os.path.join(
         output_folder, os.path.splitext(os.path.basename(video_path))[0] + ".wav"
@@ -26,7 +26,7 @@ def extract_audio(video_path:str, output_folder:str) -> str:
 
 # Function to start the extraction process
 def start_extraction(video_files: list[str], output_folder: str) -> None | list[str]:
-    
+
     if not video_files or not output_folder:
         print("Error", "Please select video files and an output folder.")
         return
@@ -36,9 +36,9 @@ def start_extraction(video_files: list[str], output_folder: str) -> None | list[
 
     # Use ThreadPoolExecutor to run multiple extractions simultaneously
     with ThreadPoolExecutor() as executor:
-        result = executor.map(
-            lambda video_file: extract_audio(video_file, output_folder), video_files
+        result: os.Iterator[str] = executor.map(
+            lambda video_file: extract_audio(video_path=video_file, output_folder=output_folder), video_files
         )
 
-    print("Success", "Audio extraction completed for files:\n", str(video_files))
+    print("Success", "Audio extraction completed for files:\n", str(object=video_files))
     return list(result)
